Update --> Django Messages 2
- Add --> In login & sign up. (Not reactive)

Fixes:
- Create --> message services. (JS)
- Add logs in django_messages.py
