+++ app/front/static/js/message.js
@@ -1,0 +1,7 @@
+import { MessageService } from "./services/message-service.js"
+
+window.displayMessages = MessageService.displayMessages
+
+document.addEventListener("DOMContentLoaded", () => {
+    MessageService.displayMessages()
+})
