+++ app/front/static/js/services/shopping-cart-service.js
@@ -40,21 +40,4 @@ export class ShoppingCartService {
         Unicorn.call("shopping-cart-counter", "update_cart_counter")
     }
 
-
-
-
-    /**
-    * Display Messages:
-    * 
-    * Use Unicorn to call methods, display django messages.
-    * @returns {void} None.
-    */
-    static displayMessages() {
-        Unicorn.call("django-messages", "add_message")
-
-        setTimeout(() => {
-            Unicorn.call("django-messages", "remove_message")
-        }, 4000)
-    }
-
 }
