+++ app/front/static/js/shopping-cart.js
@@ -3,9 +3,3 @@ import { ShoppingCartService } from "./services/shopping-cart-service.js"
 window.addToCart = ShoppingCartService.addToCart
 window.removeFromCart = ShoppingCartService.removeFromCart
 window.updateCartCounter = ShoppingCartService.updateCartCounter
-
-window.displayMessages = ShoppingCartService.displayMessages
-
-document.addEventListener("DOMContentLoaded", () => {
-    ShoppingCartService.displayMessages()
-})
