+++ app/front/templates/base.html
@@ -47,5 +47,6 @@
 </body>
 <script type="module" src="{% static 'js/base.js' %}"></script>
 <script type="module" src="{% static 'js/shopping-cart.js' %}"></script>
+<script type="module" src="{% static 'js/message.js' %}"></script>
 <script src="{% static 'js/tailwind/tailwind.config.js' %}"></script>
 </html>
