--- app/front/templates/components-base/django-messages.html
@@ -1,15 +1,0 @@
-{% if messages %}
-<div class="fixed {{ tailwind_class }}">
-    <div class="flex flex-col items-center md:items-end gap-y-4 w-full">
-        {% for message in messages %}
-        <div id="message" class="{{ message.tags }} w-full md:w-1/2 p-3 
-            text-lg md:text-xl text-white text-center rounded-sm md:rounded-md mukta-light
-            {% if message.tags == 'success' %} bg-green-600
-            {% elif message.tags == 'error' %} bg-red-600
-            {% endif %}">
-            {{ message }}
-        </div>
-        {% endfor %}
-    </div>
-</div>
-{% endif %}
