+++ app/front/templates/unicorn/django-messages.html
@@ -4,9 +4,9 @@
         <div class="fixed {{ tailwind_class }}">
             <div class="flex flex-col items-center md:items-end gap-y-4 w-full">
                 {% for message in messages_list %}
-                <div class="{{ message.level_tag }} w-full md:w-1/2 p-3 
+                <div class="messages {{ message.level_tag }} w-full md:w-1/2 p-3 
                     text-lg md:text-xl text-white text-center rounded-sm md:rounded-md mukta-light
-                    opacity-100 transition-opacity duration-300 ease-in-out
+                    opacity-100 transition-all duration-500 ease-in-out
                     {% if message.level_tag == 'success' %} bg-green-600
                     {% elif message.level_tag == 'error' %} bg-red-600
                     {% elif message.level_tag == 'info' %} bg-blue-600
